refactor: Rewrite run.go to use the KubevirtProvider
To avoid having to read each flag and return an error if its unset leverage the FlagMap, a map of flag name to FlagConfig.
a FlagConfig is the type of this flag (string, int, uint16, bool or array of string) and the option function that sets the value
of this flag on the KubevirtProvider struct.
During parsing of flags this map is being iterated on and each option gets appended to an array to later be used in the KubevirtProvider constructor.

The run method's role is now to parse the flags and pass them to the provider and just call Start.
All the floating methods in run.go are removed after being moved to the provider.

Signed-off-by: aerosouund <aerosound161@gmail.com>