Opts package (#1217)
* feat: Create K8sClient

The K8s client is a type to perform kubernetes operations on arbitrary manifests and embedded in it is the schema for all needed resources.
It includes also a test client implementation for unit tests.

Signed-off-by: aerosouund <aerosound161@gmail.com>

* feat: Introduce the CopyRemoteFile, SCP methods and Command without stdout in sshlib

CopyRemoteFile is used to copy an arbitrary file on a node or a remote entity to an io.Writer.
SCP copies to the node on a remote path.
The CommandWithNoStdout is used for commands which their output will be used later.

Signed-off-by: aerosouund <aerosound161@gmail.com>

* refactor: Reuse SSH Connection

Move the instantiation of the SSH client from happening before every command execution to be included in the
ssh client itself and only create it if it uninitialized.

Signed-off-by: aerosouund <aerosound161@gmail.com>

* refactor: Move command execution into an internal unexported method

The method takes the command to run and the io.Writer interfaces for errors and output

Signed-off-by: aerosouund <aerosound161@gmail.com>

* feat: Introduce the Opts package

This package is a go representation for each bash script that was run to configure an option for the KubeVirtCI clus
ter, where each script maps to a particular SomethingOpt type.
They all share an interface of having an Exec() error method

Signed-off-by: aerosouund <aerosound161@gmail.com>

* refactor: Introduce the previously pure bash options to nodeconfig package

Some options for the kubevirt cluster were solely dependent on the bash code, this adds them to the go code
As well as refactor the constructors for configs to the the options pattern to avoid having a constructor that takes too many args

Signed-off-by: aerosouund <aerosound161@gmail.com>

* refactor: Leverage Opts package in run.go and move pure bash flags to gocli flags

Remove the need to run any scripts in the run command and depend on the opts package.
Refactor provisionNode and provisionK8sOptions to use the options pattern instead of passing many args to the constr
uctor.
Add new flags to the gocli which were previously only in the bash code.
Remove prepareDeviceForAssignment and prepareDeviceForAssignment as they both have been moved to opts.

Signed-off-by: aerosouund <aerosound161@gmail.com>

* refactor: Delete duplicated bash code after moving it to opts

Signed-off-by: aerosouund <aerosound161@gmail.com>

* refactor: Remove the bash logic that creates things that are moved to opts

Move the checking for env variables to instead of invoking bash functions to pass flags to opts

Signed-off-by: aerosouund <aerosound161@gmail.com>

* test: Modify tests for provisionK8sOptions and provisionNode to use opts and new constructors

Signed-off-by: aerosouund <aerosound161@gmail.com>

* refactor: Modify the bump CDI and CNAO scripts to update manifests in the opts directory

Signed-off-by: aerosouund <aerosound161@gmail.com>

---------

Signed-off-by: aerosouund <aerosound161@gmail.com>