gocli: Add retry to creating and starting dnsmasq container
There is a flake that causes `make cluster-up` to fail early when trying
to create the dnsmasq container due to a port collision[1]. This looks to
be a race condition and is very difficult to reproduce.

As this happens on every lane that uses the kubevirtci virtual cluster
providers, the cumulative impact is quite high.

Adding 3 retries should help to avoid `make cluster-up` failing due to
this.

[1] https://search.ci.kubevirt.io/?search=bind%3A+address+already+in+use&maxAge=336h&context=1&type=build-log&name=&excludeName=&maxMatches=5&maxBytes=20971520&groupBy=job

Signed-off-by: Brian Carey <bcarey@redhat.com>