fix: Wait until istio cni files appear before copying them (#1262)
* fix: Wait until istio cni files appear before copying them

Signed-off-by: aerosouund <aerosound161@gmail.com>

* fix: Use CNAO v0.94

Previous version of CNAO introduced a problem with passt binding

Signed-off-by: aerosouund <aerosound161@gmail.com>

* fix: Change CNAO custom resource to not have multus in case multus is enabled

This logic is controlled by the cluster-up variable KUBEVIRT_WITH_MULTUS_V3.
If this variable is passed then it will get parsed as a gocli flag, which then gets added to the node config using the nodesconfig.WithMultus(deployMultus)
line in run.go
In this case multus is omitted from the CNAO CR and deployed using its actual manifests

Signed-off-by: aerosouund <aerosound161@gmail.com>

* fix: Remove CNAO patch from istio operator

This is the same configuration that existed prior to #1217 getting merged.
Ths addition of this lead to troubles with istio passt binding, as passt wasn't able to bind to port 22 due to insufficient permissions

Signed-off-by: aerosouund <aerosound161@gmail.com>

* test: Create test cases to assert CNAO is creating the addons config with and without Multus properly when passing the multusEnabled argument

Signed-off-by: aerosouund <aerosound161@gmail.com>

---------

Signed-off-by: aerosouund <aerosound161@gmail.com>