bug: Skip CDI deployment by default when make cluster-up is issued on slim provider
When running the slim version of provider CDI is coming up. This is because when we are setting --deploy-cdi arg, it was only set when KUBEVIRT_DEPLOY_CDI is true. We also have to set --deploy-cdi=false, when KUBEVIRT_DEPLOY_CDI=false (default) as otherwise deploy-cdi will be set to its default value true. This issue is fixed now.

Signed-off-by: chandramerla <Chandra.Merla@ibm.com>