Allow opting out of frequent etcd flushes to storage (#1266)
etcd allows opting out of WAL [0] fsyncs
(WAL records describe the commit changes and are being flushed to permanent storage very frequently)
which could significantly decrease the load on the worker pod filesystem.
This is (arguably) equivalent in terms of disasters to etcd in memory,
and has seemed to work well for other large projects [1].

As this is behind an env var, and our CI clusters are ephemeral,
there's no harm in evaluating the use of this, in hopes of squeezing
out the maximum of the resources we available to us.
(non-dedicated storage per worker)

[0] https://www.postgresql.org/docs/current/wal-intro.html
[1] https://github.com/cert-manager/cert-manager/pull/4983/commits/057ce50ace764c353ca1a90f270df4bf9b23bb64

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>