k8s: Fix permissions regardless of runtime
We pull a couple of files (kubectl, kubeconfig) from the cluster
and use them to connect to the cluster instead of whatever might
or might not be available on the local environment.

In order for this to work, the permissions have to be adjusted,
but we currently only go through that step when podman is used
as the runtime. If docker is used instead, the permissions will
be incorrect and 'make cluster-up' will fail.

We can simply drop the check and fix permissions unconditionally.

https://github.com/kubevirt/kubevirtci/issues/1293

Signed-off-by: Andrea Bolognani <abologna@redhat.com>