kind,sriov: Enable Kubevirt CPUManager FG in check-cluser-up.sh
The "check-up-kind-sriov" is turned optional and not gating because it
constantly failing, due to following tests failures:
  SRIOV VMI connected to single SRIOV network
    should have cloud-init meta_data with tagged interface and
    aligned cpus to sriov interface numa node for VMIs with dedicatedCPUs
  SRIOV VMI connected to single SRIOV network
    [test_id:3959]should create a virtual machine with sriov interface and dedicatedCPUs

The mentioned tests causing the lane to fail following removal of programmatic
skips in kubevirt/kubevirt tests kubevirt/kubevirt#13144, affecting the
mentioned tests.
Previously the mentioned tests were skipped silently (bad) and now, following
the programmatic skip removal, fail loudly.
The root cause for the failures (or previous skips) is tests depends on
Kubevirt's CPUManager feature but its not enabled at all.

Enable Kubeivrt CPUManager FG when the SR-IOV provider is tested.

Signed-off-by: Or Mergi <ormergi@redhat.com>