Kubelet drop in (#1299)
* k8s-1.30: enable kubelet drop-in config files

We can now use a proper API for customizations. See follow-up commits.

Signed-off-by: Victor Toso <victortoso@redhat.com>

* k8s-1.30: Use kubelet's drop-in config files

I've verified that in k8s-1.30, removing all options set in
KUBELET_EXTRA_ARS, would result in the following diff:

    @@ -54 +53,0 @@
    -    "kubeletCgroups": "/systemd/system.slice",
    @@ -87 +86 @@
    -    "failSwapOn": false,
    +    "failSwapOn": true,

I've moved both kubeletCgroups and failSwapOn into kubelet's drop-in
config file and verified that it works and config is set.

I assume that --cgroup-driver=systemd is already the default in centos
stream or being set somewhere else, but to avoid regressions I've
included in the config file as well.

I don't know why --runtime-cgroups is not being shown in the node's
configz but it is the only command line option for the initial 4 that
is not marked as DEPRECATED in favor of Config file alternative, so
let's keep it as command line option, together with --config-dir.

Signed-off-by: Victor Toso <victortoso@redhat.com>

* k8s-1.31: enable and use kubelet's drop-in config files

Apply similar changes done to k8s-1.30, see commits:
 * "k8s-1.30: enable kubelet drop-in config files"
 * "k8s-1.30: Use kubelet's drop-in config files"

Signed-off-by: Victor Toso <victortoso@redhat.com>

* k8s-1.32: enable and use kubelet's drop-in config files

Apply similar changes done to k8s-1.30, see commits:
 * "k8s-1.30: enable kubelet drop-in config files"
 * "k8s-1.30: Use kubelet's drop-in config files"

With this, future k8s will have this enabled too.

Signed-off-by: Victor Toso <victortoso@redhat.com>

---------

Signed-off-by: Victor Toso <victortoso@redhat.com>