fix: gocli startup in s390x architecture (#1354)
The gocli container image is multi-architecture compatible;
however, it attempts to download the k8s-1.30 image, which is not
available for the s390x architecture. The k8s-1.30-slim image, on
the other hand, is compatible. To enable support for s390x, we can
enforce the use of the --slim parameter for this architecture.

Signed-off-by: Nestor Acuna Blanco <nestor.acuan@ibm.com>
Co-authored-by: Nestor Acuna Blanco <nestor.acuan@ibm.com>