kind-sriov: Bump SR-IOV CNI and device-plugin (#1365)
* kind-sriov: Bump SR-IOV CNI and device-plugin

Signed-off-by: Or Mergi <ormergi@redhat.com>

* kind-sriov: Update sriov-dp manifest

According to example [1]

Previous, the manifest included additional architecture aware daemon-sets.
The manifest now utilize the container manifest [2], making the additional
daemon-sets redundant; the container runtime will pull the relevant image
according to the host architecture.
In addition, the SR-IOV provider does not support arm or ppc64el arch.

[1] https://github.com/k8snetworkplumbingwg/sriov-network-device-plugin/blob/v3.9.0/deployments/sriovdp-daemonset.yaml
[2] https://github.com/k8snetworkplumbingwg/sriov-network-device-plugin/pull/543/commits/903b5973cb7e6fc26e38a08801853adeb5c2c502

Signed-off-by: Or Mergi <ormergi@redhat.com>

* kind-sriov: Update sriov-cni manifest

According to example [1]

[1] https://github.com/k8snetworkplumbingwg/sriov-cni/blob/v2.9.0/images/sriov-cni-daemonset.yaml

Signed-off-by: Or Mergi <ormergi@redhat.com>

* kind-sriov,kustomization: Use patches keyword

The Kustomization field 'patchJson6902' is deprecated, 'patches'
field replaces it.

Signed-off-by: Or Mergi <ormergi@redhat.com>

* kind-sriov,node-selector patch: Init the node-selector field

In case the node-selector field is omitted in the manifest the patch
will fail.

Change the patch to initialize the node-selector field before the label
patch.

Add test patch to ensure the expected node-selector is in place.

Signed-off-by: Or Mergi <ormergi@redhat.com>

---------

Signed-off-by: Or Mergi <ormergi@redhat.com>