fix: segmentation fault when using podman
The ephemeral-provider-common.sh script reveals several issues
related to Podman, particularly when dumping file content to
stdout. The CLI scp command is affected; however, by replacing the
string "- >" with the destination folder itself, the command
functions correctly.

Additionally, it's important to note that the destination folder
differs between Podman and Docker due to changes made in
ephemeral_provider_common.sh on line 57.

As an alternative, we could consider making this workaround more
generic by eliminating the use of "- >" altogether.

Signed-off-by: Nestor Acuna Blanco <nestor.acuna@ibm.com>