fix,pre-pull-images: move fetch-images.sh to manual
since the manifests are not inside the cluster deploy directory any more
they are not being copied into the directory any more, thus the
fetch-images.sh will not find any images.

Therefore we move the fetch-images.sh into the main provision/k8s folder
and replace the per version with a pre-pull-images text file.

This file needs to get updated whenever the manifests inside either the
k8s version folder or the manifests inside the gocli change.

Thus we also create a script update-pre-pull-images.sh and call that
whenever the cluster is checked at first.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>