Fix: Bypass Provision Manager for s390x kubevirtci publish
As of now, we are updating kubevirt-prow/release/kubevirt/kubevirtci/latest file after amd64 images are published, independent of s390x image publish failure. Provision manager assumes s390x images are available with the tag in the latest file and tries to retag them when no diffs between the current code and the tag. But this fails, when no s390x image exists with tag in the latest file

So, this workaround is required till we make tag in latest file be updated after s390x image is published as well.

Signed-off-by: chandramerla <Chandra.Merla@ibm.com>