fix: remove aaq and kwok, disable kwok install
We don't want to include aaq and kwok into the pre-pull, so we are
removing them from the pre-pull lists.

Also we are externalizing the single exclude patterns into a file, since we
already have a grep with exclude pattern in the script.
There we are also adding the ones for kwok and aaq.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>