fix: segmentation fault in podman (#1387)
* fix: segmentation fault in podman

For more information, please refer to:
https://github.com/containers/conmon/issues/315

The decision to create an additional function instead of using a
variable is based on the behavior of shell parameter expansion.
Specifically, defining a variable with a character such as '>'
will result in the Bash shell adding quotes before executing the
command.

Signed-off-by: Nestor Acuna Blanco <nestor.acuna@ibm.com>

* common: fix permission denied with kubectl

Use the kubectl from the k8s provisioner and set the expected
permissions on kubectl and kubeconfig.

Co-authored-by: Nestor Acuna Blanco <nestor.acuna@ibm.com>
Signed-off-by: Alice Frosi <afrosi@redhat.com>

---------

Signed-off-by: Nestor Acuna Blanco <nestor.acuna@ibm.com>
Signed-off-by: Alice Frosi <afrosi@redhat.com>
Co-authored-by: Alice Frosi <afrosi@redhat.com>