In Provision Manager, call "git fetch --tags" command to avoid git diff fail (#1376)
* Add git fetch tags command to avoid git diff fail

Without change, in provision manager while doing git diff to a tag obtained from latest file in gcs, if that tag is not present on local clone, we are exiting with a message to run git fetch by the user. Now with this change, we are once 'trying' to run the git fetch tags command after add upstream remote, before asking the user to do that.

Signed-off-by: chandramerla <Chandra.Merla@ibm.com>

* Removed branch fetch during git fetch tags command

Signed-off-by: chandramerla <Chandra.Merla@ibm.com>

---------

Signed-off-by: chandramerla <Chandra.Merla@ibm.com>