vm-image: Upgrade fedora-with-test-tooling to Fedora 39
Fedora 35 has been EOL for a very long time and it doesn't make sense to
continue to test against it.

Updating to Fedora 39 (also EOL but more recent)

This has been tested against the kubevirt test suite and all tests are
passing apart from some ipbv6 only tests which have a fix on the way.

[1] https://github.com/kubevirt/kubevirt/pull/13322
[2] https://github.com/kubevirt/kubevirt/pull/14275

Signed-off-by: Brian Carey <bcarey@redhat.com>