cluster-prv, gocli, istio: Bump istio to 1.24.4 (#1411)
* cluster-prv, k8s, base-img: Bump base image

Signed-off-by: Nir Dothan <ndothan@redhat.com>

* cluster-prv, gocli, istio: Bump istio to 1.24.4

Istio fix of bug https://github.com/istio/istio/issues/54820
which was critical to kubevirt, was backported to 1.24.4.
The WA hack, of injecting the istio-cni annotation is no longer
required and thus removed from istio-operator-with-cnao.cr.yaml.
In addition, values.cni.provider: multus
was added, it causes auto generation of NAD in default ns.

Signed-off-by: Nir Dothan <ndothan@redhat.com>

---------

Signed-off-by: Nir Dothan <ndothan@redhat.com>