Export CDI upload proxy lowerband port in CI clusters
Currently CI clusters already expose the upload proxy port which is
later used for setting up a node-port.

Because the port belongs to the dynamic node-port allocation
range (ala upper-band) by default[1] it makes it so that node-port can
be claimed by a dynamic service which in turns causes a conflict once
the upload proxy service is attempted to be created.

This commit adds a temporary UploadProxyLowerBand port from the static
range to fix that.

[1] https://kubernetes.io/blog/2023/05/11/nodeport-dynamic-and-static-allocation/#default-range-30000-32767

Signed-off-by: Adi Aloni <aaloni@redhat.com>