istio: Remove Multus settings from the default Istio config (#1422)
The default Istio config 'istio-operator.cr.yaml' is intended to be used
when Multus is not installed in the cluster.
Multus depends on the NetworkAttachmentDefinition CRD to be installed.

Currently the default Istio config is invalid, because it specify
setting that installs the istio-cni NetworkAttachmentDefinition CR,
where it should be used when no Multus is installed.
For setups that have Multus installed, the
'istio-operator-with-cnao.cr.yaml' should be used.

Remove Multus related settings from the default Istio config.

Signed-off-by: Or Mergi <ormergi@redhat.com>