fix: use pre-generated image lists for pre-pulling images (#1374)
* fix,pre-pull-images: move fetch-images.sh to manual

since the manifests are not inside the cluster deploy directory any more
they are not being copied into the directory any more, thus the
fetch-images.sh will not find any images.

Therefore we move the fetch-images.sh into the main provision/k8s folder
and replace the per version with a pre-pull-images text file.

This file needs to get updated whenever the manifests inside either the
k8s version folder or the manifests inside the gocli change.

Thus we also create a script update-pre-pull-images.sh and call that
whenever the cluster is checked at first.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* fix: remove aaq and kwok, disable kwok install

We don't want to include aaq and kwok into the pre-pull, so we are
removing them from the pre-pull lists.

Also we are externalizing the single exclude patterns into a file, since we
already have a grep with exclude pattern in the script.
There we are also adding the ones for kwok and aaq.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* fix: make image names w/o registry explicit

We replace all implicitly docker.io images, i.e. those that omit the
registry name (i.e. grafana/grafana) by an explicit addition of the
docker.io prefix.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* fix: update cephcsi version

Changes:
* update 1.33 with changes from PR
* increase memory on check-cluster-up.sh to run tests in parallel

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* fix: update pre-pull-images

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

* chore(check-cluster-up): size down single node check

Signed-off-by: Daniel Hiller <dhiller@redhat.com>

---------

Signed-off-by: Daniel Hiller <dhiller@redhat.com>