vm based providers: Timeout if node hangs
Sometimes the node hang, for unknown reason,
lets set timeout about it, instead waiting for 3h.

Next time will be to add screen /dev/pts/1 to see what happens in the VM,
when it hangs.

We had it at the past, and it gone https://github.com/kubevirt/kubevirtci/pull/1049
Now it returned.

Signed-off-by: Or Shoval <oshoval@redhat.com>