flannel: Set MTU to 1480
Flannel creates flannel.1, flannel-v6.1 VXLAN tunnel interfaces.
Since each of them have only one ip family, the kernel sets
different MTUs which are the host MTU - 50 / MTU - 70 depending on the family.

This isn't the same behavior we have on Calico / OVN-k8s, so tests
which check unfragmented packets with MTU size (for example on kubevirt),
will fail on IPv6 since they depend on pod MTU size which is the same for both
families, while the node has an additional tunnel with different size for each family.

Setting the MTU on the backend, allows to have both tunnels and pod
with the same MTU, which is MTU - 50.

Note:
We can't set MTU 1500 as the default, because it will result in the
original behavior.

Kubevirt test that failed without it (IPv6 one):
https://github.com/kubevirt/kubevirt/blob/2fbb6145135695cc8f320c4197891a504b4bfb9c/tests/network/vmi_networking.go#L786

$ ip link show flannel.1
4: flannel.1: ... mtu 1450 ...

$ ip link show flannel-v6.1
5: flannel-v6.1: ... mtu 1430 ...

Signed-off-by: Or Shoval <oshoval@redhat.com>