vm based: Fix Flannel max pods per node
controllerManager change was also needed, beside Flannel params.
Follow-up fix for https://github.com/kubevirt/kubevirtci/pull/1557

Sumup:
We are testing on CNAO lifecycle lane install / uninstall of components, one of them is Multus.
Multus is mandatory in order to propagate the CmdDel command from kubectl to host-local plugin,
since multus is down when some pods are deleted, CmdDel are lost,
hence there is ip leakage and the pool is depleted, because the default pool is a very small /24.
Flannel doesn't have GC for ips.

Hence this PR fix that the pool will be /20 (both Flannel and k8s changes were required, this PR adds the missing one).
This fix is enough, since the bug happens just on lifecycle which checks a corner case that affects Flannel.

Signed-off-by: Or Shoval <oshoval@redhat.com>